manga/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
idna==3.7
joblib==1.4.2
kombu==5.3.7
numpy==1.26.4
pandas==2.2.2
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
